wlan: Avoid dropping packets if data packet filter is set
Currently, when MSDU reorder info exists for a tid of station, then
host last Rx packet sequence number SSN goes out of sync with the
hardware. As a result data packets other than the filtered packets
in firmware which arrive on a particular tid to host will get dropped.
In this change avoid dropping the packet if data packet filter is set.

Change-Id: I4128e3be118afd76fd4abacc45f6ceb0f865ba5e
CRs-Fixed: 1088971
Signed-off-by: FireLord1 <firelord.xda@gmail.com>